+++ src/data/recorder.ts
@@ -184,6 +184,18 @@ export const statisticsHaveType = (
   type: StatisticType
 ) => stats.some((stat) => stat[type] !== null);
 
+const mean_stat_types: readonly StatisticType[] = ["mean", "min", "max"];
+const sum_stat_types: readonly StatisticType[] = ["sum"];
+
+export const statisticsMetaHasType = (
+  metadata: StatisticsMetaData,
+  type: StatisticType
+) => {
+  if (mean_stat_types.includes(type) && metadata.has_mean) return true;
+  if (sum_stat_types.includes(type) && metadata.has_sum) return true;
+  return false;
+};
+
 export const adjustStatisticsSum = (
   hass: HomeAssistant,
   statistic_id: string,
